feat: 🎸 adjust buy
