+++ packages/mobile-app-did/js/pages/Buy/BuyHome/components/BuyForm/index.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import BuyMainForm from './main';
-import BuyTestForm from './test';
 import { useIsMainnet } from '@portkey-wallet/hooks/hooks-ca/network';
 
 export default function BuyForm() {
   const isMainNetwork = useIsMainnet();
-  return isMainNetwork ? <BuyMainForm /> : <BuyTestForm />;
+  return isMainNetwork ? <BuyMainForm /> : null;
 }
