+++ packages/mobile-app-did/js/pages/Buy/hooks.tsx
@@ -11,79 +11,6 @@ import { INIT_HAS_ERROR, INIT_NONE_ERROR } from 'constants/common';
 import isEqual from 'lodash/isEqual';
 import { useIsFocused } from '@react-navigation/native';
 
-export const useTestAmountPrice = (
-  // receiveAmount: number,
-  // fiat?: FiatType,
-  // token?: CryptoItemType,
-  _receiveAmount: number,
-  _fiat?: FiatType,
-  _token?: CryptoItemType,
-  initialAmount?: string,
-) => {
-  // const [amount, setAmount] = useState<string>(initialAmount || '30.73');
-  const [amount] = useState<string>(initialAmount || '30.73');
-  // const [rate, setRate] = useState<string>('0.31');
-  const [rate] = useState<string>('0.31');
-  const [rateRefreshTime, setRateRefreshTime] = useState<number>(MAX_REFRESH_TIME);
-
-  const rateRefreshTimeRef = useRef(MAX_REFRESH_TIME);
-  const refreshReceiveRef = useRef<() => void>();
-  const refreshReceiveTimerRef = useRef<NodeJS.Timer>();
-
-  const clearRefreshReceive = useCallback(() => {
-    refreshReceiveTimerRef.current && clearInterval(refreshReceiveTimerRef.current);
-    refreshReceiveTimerRef.current = undefined;
-  }, []);
-
-  const registerRefreshReceive = useCallback(() => {
-    clearRefreshReceive();
-
-    refreshReceiveRef.current?.();
-    rateRefreshTimeRef.current = MAX_REFRESH_TIME;
-    setRateRefreshTime(MAX_REFRESH_TIME);
-
-    const timer = setInterval(() => {
-      rateRefreshTimeRef.current = --rateRefreshTimeRef.current;
-      if (rateRefreshTimeRef.current === 0) {
-        refreshReceiveRef.current?.();
-        rateRefreshTimeRef.current = MAX_REFRESH_TIME;
-      }
-      setRateRefreshTime(rateRefreshTimeRef.current);
-    }, 1000);
-
-    refreshReceiveTimerRef.current = timer;
-  }, [clearRefreshReceive]);
-
-  useEffectOnce(() => {
-    registerRefreshReceive();
-    return clearRefreshReceive;
-  });
-
-  const refreshReceive = useCallback(async () => {
-    return;
-    // if (fiat === undefined || token === undefined) return;
-
-    // try {
-    //   const rst = await getOrderQuote({
-    //     crypto: token.crypto,
-    //     network: token.network,
-    //     fiat: fiat.currency,
-    //     country: fiat.country,
-    //     amount: '100',
-    //     side: 'BUY',
-    //   });
-
-    //   setRate(formatAmountShow(rst.cryptoPrice, 2));
-    //   setAmount(formatAmountShow(receiveAmount * Number(rst.cryptoPrice), 2));
-    // } catch (error) {
-    //   console.log('error', error);
-    // }
-  }, []);
-  refreshReceiveRef.current = refreshReceive;
-
-  return { amount, rate, rateRefreshTime };
-};
-
 export const useReceive = (
   type: TypeEnum,
   amount: string,
